stopping mid-hmmmmmmm (see file-io.c)
So I started working on handle.c and got quite a way until it came to
what is theoretically in the starthost() function.

What I want is for the caller to execute edb_open and if there's no
host process associated with the file, said host process is started
implicitly and subsequent calls from seperate threads and processes
will not have to also start their own hosts.

as convienet as that is; im starting to think that more explicit
action needs to be taken on the caller side. as if I fork(2), and
use the child process to operate as the host, this will also fork
all the code of the caller, and we don't want that.

using fork(2) led me to use ofd locks. I think a better way of
doing this would be execve though?

hmmmmmmmmmmm