more work on the db. focused on the host function.
move a little bit of code around. stopped in edb_host.

back to being in the status of "just keep coding"... the only major hole in the structure i see how the page buffer will work. next time i need to think about putting this buffer inside of the context of the file maybe?

need to also keep in mind chached pages.