removed the -load and -read relationship on the main interface in leu of -copy
I did this when looking of the usefulness of -load and -read vs -copy. And -copy will pretty much win every time. It would be nice to have something along the lines of -writelock and -readlock someday but for right now we'll stick with -copy.

This also allows easier engineering with shared memory.