to address the transfer problem of arbitrary size information between host and handle I put in a "jobtransfer buffer"
This involved putting a mutex on each job as well as a futex so that either side of the jobs can stream data to and from.

This stream is also configurable with the use of an intermediate buffer. I think a better way of doing this is to NOT use an intermediatary buffer, however that would involve a more complicated interface that allows for allocation of buffers in the shared space per the handle's will. Which is against the "static allocation" doctrine.