First draft of rowoffset_lookup
Bit change is that lookup pages can no longer be in straits. That sort of engineering would be too complex and not worth it. This made things a lot easier. Leaf nodes are stil. in straits.

Also... see agf changes.

Not as hard, but my brain hurt a little bit. Need to find out CREATING.