edba_u_lookupdeepright: BROKEN COMMIT. Implemented a large chunk of logic.
I'm commiting this now to save my progress in the cloud. I left a good amount of todos and concerning typing. Still much to do in the function.

I also got the locking specification in my notebook and need to put it in locking.org. But as you can see from one of my todo's, the locking still have some thinking needed.