edba: structure editing BROKEN COMMIT
Started to implement the structure creation/deleting. And of course I ran into an oversight in engineering (being that structures was THE first thing I engineered and haven't touched sense).

The problem is space managment with deleted structures. See TODOs in edb.org and edba.h.

Some things that were decided and made progress on:

 - Structure flags
 - Some documentation clarificatoins.
 - More working variables jammed in edba_handle_t :).