spec: a really elegant solution to solving the structure spacing problem.
Well it's happened. The database is starting to design itself. All I did was reuse the edbp_object spec into the edbp_struct with some execptional behaviour.

This is iconic. OidaDB's design is now a "recursive" design, and by defintion (the only truth is one that proves itself), this makes the design TRUE.

Anyone who reads the above that isn't me is going to cringe. But thats okay. I'm me. And me is right. Which means I only need to listen to me.