edba-struct: implemented functions
edbw is now exposed the a simular workflow as edba-objects in the sense that the structure must be opened before deleted. This was done because I want the structure-page-lookup logic to exist only in the create functions. Plus the design is simiular to what I have already.

Also elaborated some possible errors with edba_u_dynamicdelete.

This resulted in various changes in a few places. Also updating the locking.org.