odbm: arch juices are finally flowing!
See elements_u.h... notice what I'm doing. I'm making "parent classes" in the utility header and I'm going to "derive" them in child classes that are exposed by elements.h

Also have to keep a close eye on figma for this so I know what to call things in advance.

TODO: need to define the namespace relationships of the odbm domain.