edbp: passing multithreaded tests.
We're running about 8% faster than the OS's PRA. Play with the config variables for more scientific deductions. I'm not commiting the PRA code at this time because it needs to be refined and not in the scope of this card.

I went HARD with this test: 150k pages (~1G), 16 threads performing 10k test each and a cache of 4096. This took about 6 seconds, but the results table is very promising. I'll tone it back for future (simpler) tests to be faster.

edbp is WELL TESTED.