oidadb.h: just some thoughts on how the interface should work... checking off some todos.
I also theorized the idea of "polled responses" so that when it comes to job compression, a single handle can benifit from job compression by submitting hundreds of jobs before even worrying about the results.

Lastly, see the new reference in edb.org. Its definitely provided some insperation for the aformentioned idea, if not an improvement.