edbs-jobs: job communication redesign
Equiped with a better idea how jobs will work, I've been having to design how the read/write functions should work with the buffer.

What is complicated here is that a bi-direction pipe isn't as easy as simply read and writes. Some rules must be made to make things way less complicated than it must be.

I've drafted up my first thoughts on how they work. But I need to dive in and try to implement them and see what sort of resistance I get. But we're leaving the analogy of read(2)/write(2).

Another though I must have is the possibility of:

 - a handle submitting multiple jobs before it even try to read the response from the first one
 - a handle streaming the results from 1 job into the other. (this is actually a really interesting concept).