edbs: testing returning success. Had to make a few threading adjustments.
So there was a few tiny typos here and there. Those were quick to find.

A more challenging thing is the communication mutexes for the transfer buffers. I had a logical pitfall thinking 2 futexes where enough. They were not. More futexes never hurt.

The BIG hair-ripping one resulted in the conception of line 43 in edbs.h. "edbs_host_close". The absence of this function was an absolute nightmare and I didn't even know it. Basically, I needed a method to use to let the threads in the test to actually join before I could call free.

Another small discovery was when the host had shut down before the handles. This messed with the mutexes in the stray handles. I solved this by added "Last man out hit the lights" logic so that it was everyones responsibility to clean whoever disconnected last.