Designed polling telemtry.
This was better than callbacks because now I can use them with other processes. Besides... polling is almost always better then callbacks.

I DID SCRAP THE "FLAWLESS" IDEA. I needed to do polling for innerproc, but, to make such a thing flawless without the possibility of EDB_EMISSED was... hard and confusing. It would be better if I just did the lossy poll method to prepare for the events rather then spend so much time on a feature that's not even that important.

Defintions coming next.