edba-testing: upping rows to 20k revealed a "bug" that wasn't really a bug but more so an error that isn't really an error.
The EDB_FUCKUPS had an if statement that was assumed to be an error due to bad logic. But this reality of "bad logic" turned out to be just a very small chance that it could happen and thus should be handled as such. I just removed the if statement from fuckups and placed it before the lock.

See the large comment in the if statement.