edba-testing: bugs squashed (padding drama)
See edbd-init.c: that was the original cause of the problem I was having. I used the wrong structure.

But in going down the rabit hole, I realized that padding was being added to odb_spec_struct_full_t... which is a rare case that it must be packed to obey the schematics of an object page. I added a check in odbspec.c for future reference.

If this wasn't packed, then it'd add 4 bytes between the flags and dypointer. Those 4 bytes are NOT supposed to be there in a normal object page.