edba-testing: Testing went surprisingly well. Though I'd like to introduce delete statements into the MT enviroment.
After playing around with the settings, our biggest bottlenecks are setting OFD locks (note: NOT waiting for other threads, but setting them in general). It seems that its an expensive operation that requires an mutex and thus slows all the threads down.

The edbp_start and edbp_finish seem to be also causing bottle necks between threads. Page faults are not a problem.

But thankfully, looks like we're seeing faster times in the MT environment overall. The biggest hurdle is edbl's setting of locks, not the actual collision of them.