manual: working through it even more.
Right now I'm focusing on odbh_jobdesc_t.org. Going through this, its really causing me to check my work on a lot of the design desicions I've been rolling with. Its not really complete. But, I'd say this is the 3rd/4rd draft of what the frontend will look like and I feel like this may be *it* (for v1).

I also added the "tradesecrets" tag so I can document my thoughts that shouldn't be exposed later.

My next commit will be dumping a lot of shit out of the worktree for vaiours reasons.