Cleaning working tree: refactored odb_createparms -> *_t
as well as some comment cleaning in odbadb.h and some verbige correction in liboridadb.org.

I'm doing this cleaning becuase it has occoured to me that for the past several weeks I have been pretty aimless. Though I'm making progress here-and-there, I'm not really focused with my work.

So I hopped back on the project board and got my act back together. The new focus is GETTING BUILD 1. The thing build 1 is missing now is the handle-hosting boot up sequence.

Build 1 isn't really a big milestone, and is not near v1.0.0, but it is a small little target I'd like to hit as to progress the final wave of engineering that needs to be done. I'll probably take this opertunity to also go back and forth with the man/ too to clean that up.