Invented the seperate ides of ELMTRASH and ELMDEL
See locking.org for the literal change.

The reason for this is because I need a way to allow =odbh_index= to tell the user if they've downloaded the entire index or just part of it. There should be commit not far after this with updates to odbh_index.org with details in which this seperation is used.