edbh: defined odb_index
The big TODO on this one is the one you see in edbh_u.h... right now it only downloads the index on the first call and never again. Need to define odbh_poll to detect a re-download.

It also downloads the whole index which can be inefficient. Should only need to download pages as we need them sense the index can get up to gigabytes in size.

I'm going to now defined odb_structs to keep everything moving on.