edbd: fixed oversight on how fcntl OFD locks work.
I thought I could get the host PID to the handle by means of using fcntl with the locks and checking the l_pid ability. But, turns out according to fcntl(2):

"If the conflicting lock is an open file description lock, then l_pid is set to -1."

So I'll be stuffing the pid in the file header. Pretty easy task, as I was planning on doing this anyways.