edb.org: inventing the idea of "Structure Version"
Start by reading the notes found in edb.org. All other files you see in this commit are in persuit of that change.

I got this idea from going back through edbw and the edge case kinda reveiled itself. But it was a good edge case because I came up with the idea ODB-95.

(Also ODB_EUSER: new error, crept in)