edbw-objs.c: defiend jupdate and jselect
 - changed objc to a unit32_t because made no sense to have it as a uint64_t
 - Let a -1 objc symbolize an error.

I had to make some new (undefeind) edba functions though to allow me to edit entire pages at a time. It breaks the flow a little bit with the idea of the edba_pageadvance but it worked out pretty well.

Note to self: desperately need to get routines soon. This jupdate is shit. jselect is pretty nice though.