jobflush: elegant solution
Well depends on your definition of elegant. Basically, I had to clarify some of the existing rules to include the possibiilty of a client-buffer.

I also had to modify some code to allow the returning of ODB_EEOF, see that error for a description on what I changed.

This basically will build in my theorized edbs_jobflush in edbs_jobterm.