implemented some block commands
Consolidated some of the blocks functions together because it was more efficient that way.

Had to readjust my plans how how I place the blocks in buffers... I was hoping to mmap(MAP_PRIVATE) right into the buffers but then I saw this in mmap(2):

 "It is  unspecified  whether changes made to the file after the mmap() call are visible in the mapped region."

Which doesn't work for us, so I opted to use read(2) for now. see the later comment in pages-blocks.c:73

 - defined
 - defined group_loadg, group_unloadg