massive changes to the odbv1 specification.
basically, I unintentionally made a file system in that last draft. So I removed the entire idea of journalling as well as using "swap" pages in leu of just operating out of memory and trusting the real FS to do the work of making sure everything is atomic.

This simplified a lot. There's still some inconsistencies in there but for the most part everything is figured out.

Also slid in some minor changes to pages.h and buffers.h as well as cmakelist update.